fix: moved lido oracle testing in one place
